requirements.txt.txt
streamlit
PyPDF2
python-dotenv

langchain-text-splitters
langchain-community
langchain-openai
langchain-core

faiss-cpu
